info: show the series history log if available
Here we avoid the Common.expectJson logic which
somehow imposes a String in a position where we can't
handle it (the elm compiler, as nice as it is, cannot
pinpoint the origin of this contraint).

Hence we do a more gradual thing:
* get the raw payload out of the Http pipeline
* process it

It turns out debugging this is quite easier.