Reset offset and delete from browser cache
A modification of the offset on a specific horizon is only relevant to that particular horizon. When you select a horizon and adjust the offset, it is necessary to reset it to 0 when choosing a new horizon. Since the offset is reset to 0 with each change of horizon, it is also desirable not to store it in the browser cache.