Fix select horizon plot data
During the application startup, getplot was previously called twice: once in the init function and another time in the sub routine. However, it is now necessary for it to be first invoked by the sub routine. If no values are stored in the browser cache, it should be called through Cmd Msg.

Additionally, the calculation of bounds was unnecessarily performed in the updateHorizon function.