Reshape horizon
We now incorporate the JSON sent by the browser cache for the value of the horizon into the model. We are now working with a Maybe String, which is more suitable for handling horizons.
As a result, we remove the "All" value from our horizon dictionary, as this option now corresponds to the case where the horizon has the value Nothing.