Structural modification
The configuration of the "offset" button is revised: the dates corresponding to the boundaries of the selected interval are now displayed on either side of the horizon selection form.
The HTML code structure is also modified, with the integration of revision dates and the "offset" button into a table (grid), aiming to simplify and enhance the layout.