Save Button Addition
A button is added next to the summary table of modifications, allowing users to save the new values in the database via a PATCH request to the "state" API endpoint. If the data is successfully saved, the main table updates and the modified values appear in red. The secondary table is replaced by a message indicating the success of the recording. If the operation fails, the secondary table remains, displaying a message inviting the user to try again.