Model Update
To integrate the cell paste code from Excel, we need to adjust our model to streamline table updates while the user edits data. We're adding two new parameters, "editedValue" and "index", to the record. Each value in our time series needs to be indexed to find the next n dates from an entry date. In Elm, calculating dates based on an original date and frequency isn't currently possible, so we're employing this method. The currently edited value will be stored in "editedValue".