Copy paste action
We apply the code generating pasting from Excel to our table. When a user selects a cell to paste one or more values, it fills in by placing the first value of the paste in the selected cell and the next n values in the n cells below the initial cell. If the cells are not empty, the previous values are automatically overwritten.