Insertion date Check
Before saving freshly modified data, a check of the insertion date is required. When a user makes a modification and saves it, it's associated with an insertion date, determining the latest version of the series. If a user makes changes while another user is also making modifications and saves them simultaneously, the first user may not have the updated table. Hence, before allowing the user to save their modifications, we verify that the insertion date of the series they are modifying matches the latest one saved on the server.