Timezone Option
We introduce a new parameter called "timezone" to allow the user to select the one that suits them best. When the user changes the time zone, a new request is sent to the server, and the modifications are then displayed. Only two values are available (UTC and CET), but the list can be expanded. By default, UTC is selected.