Add inferred_freq option
We are adding the "inferred_freq" parameter. It is presented as a button and allows detecting and highlighting any gaps between two values in the series. When the user activates this button, a request is sent to the server, and the data is updated. This enables the user to easily identify and correct missing values. Additionally, at the boundaries of the series, if there are not enough values for a given period, inferred_freq fills the gaps by adding "NaN" values until the requested limit.