Serie cache locker
When the user is working on a "formula" type series, they have the option to cache it to avoid long loading times, especially when it contains many points. This is done by activating a button. As this action may take several seconds, the button is disabled during execution to prevent further clicks, and a "Loading..." message informs the user that the action is in progress. Once completed, the button becomes active again.