Graph and table synchronization
Until now, when the user navigated on the Plotly-displayed graph and zoomed in, the data did not update in the table above. Thus, we've enabled synchronization between the displayed interval on the graph and the data in the table. This significantly simplifies accessing the points the user wants to modify.