Horizon widget CSS transition
The horizontal widget is exportable to all views of the page where it is needed. We reorganized the styling for displaying this widget by creating a CSS file. However, this file is not compatible with the Bootstrap framework used on the pages. Therefore, we have temporarily created a horizontal widget managed by Bootstrap rather than CSS. This is temporary, and eventually, it will be removed, and we will import the widget linked to the CSS.