Offset button unblock
We remove the button blocker from the horizon widget, which prevents users from sending requests when the displayed horizon is outside the series bounds. Sometimes, when the series loads, no data is displayed if the requested horizon (by default 2 weeks before and after the request date) is outside the series limits. Thus, users should be allowed to manually recenter using the widget arrows.