Histroy graph queries
Using the list of insertion dates, we can send a new request to display the curve's state within the selected interval as it was at the time of insertion. If there are more than 70 insertion dates, to prevent overloading the server with requests, we limit and select only the first 70.