Horizon management
We introduce a message to handle cases where the user prefers to display the horizon predefined by the widget rather than manually selecting a date range by zooming in on the main graph. To avoid issues related to excessive server requests, we do not synchronize the horizon widget with the historical curve graph.