formulas: create request
To retrieve the list of formulas, we perform a GET request on the
"find" API endpoint. This request accepts several parameters,
including "query," which allows us to target the desired formulas.

In our case, we set "query" to "(by.formula)" in the formula language
and set "meta" to True to access not only the names but also the
formulas themselves.