tsinfo: ensure integrity between view and request
When a big number of revision are requested, one could
change its zoom paramter, and the old request does
not make sense anymore. With a simple monotonic counter,
we can discard the request out of date.

It can lead to a long waiting time though: we should also
be able to purge the fetch pool handled by the browseer