search: reimplement the search using the .find API
This makes the whole thing cleaner and simpler, also always up to date.

It also fixes bugs related to federated instances (we didn't have
their metadata).

It is very responsive on a 20k series dataset.

I a future step, we may want to not load everything since we in
practice only show 1000 items. Find provides a handy `limit`
parameter.

Though we have lost the ability to query only on metadata values. We
may want to bring that back (we need a `by.metavalue` filter in
tshistory).

closes #24