quickview: cancel pending series requests when launching news
When the call context change (change in horizon by the user),
we already had a mechanism to discard the the obsolete ones.
Now, the are also canceled.

For some reason, does not work with group: to be investigated

It seems that the "cancel" method dos not work with the falsk
wsgi launcher: to be tested on a proper infra