folders: ensure that all errors are counted in the transactions
It assure that the source and destination folders are correctly
server-updated after a series deplacement, even in case of error.
An more sophisticated approch would imply to retry the errors.

Note: the error appear spontaneously when moving more that 1300
series at once and leads to browser errors that does not process
all the fetch query
(Failed to load resource: net::ERR_INSUFFICIENT_RESOURCES)