tseditor: test still closer to actual data-flow
The "editable" attribute is mainly driven by the type of the payload:
a Scalar is not editable while a Complex is.
Hence, the test set the primary as Complex and the other types as Scalar

The results are fairly coherent, with the exception of cells produce
by "missing" values in a outer-merge. This aspect will highlighted
(and corrected) in the following test changes