q-editor: implement client side the operators by.at-path and by.without-path
For now, the parameter "children" is mandatory:
the logic for optional parameters with default values
seems to be lacking for now.

In term of implementation, we expand the ad-hoc parsing functions.
This operator needs a string and a bool, hence we defined
"strandbool" that parse a string and a boolean.

A futur ovehaul of this code should allow it to be more generic.

related #214