quickview: hide link to quickview while editing
This is a copout, but it is easy: since we use the basket name
as reference in the permalink, only a saved basket can be linked.
A more complete solution would be to use the query itself in the
permalink if we are in an unsaved.
It would need a more dedicated model, though.

related #203